printf like scimark c

+++ jnt/scimark2/MonteCarlo.java
@@ -33,7 +33,7 @@ public class MonteCarlo
 {
 	final static int SEED = 113;
 
-	public static final double num_flops(int Num_samples)
+	public static final double num_flops(long Num_samples)
 	{
 		// 3 flops in x^2+y^2 and 1 flop in random routine
 
@@ -43,7 +43,7 @@ public class MonteCarlo
 
 	
 
-	public static final double integrate(int Num_samples)
+	public static final double integrate(long Num_samples)
 	{
 
 		Random R = new Random(SEED);
